# public/swagger.yml
openapi: '3.1.0'
info:
  title: MailService
  version: '1.0.0'
  description: MailService REST endpoints
  contact:
    email: uwu@zhiyan114.com
    url: https://github.com/zhiyan114/email-queue
servers:
  - url: https://api.mail.zhiyan114.com
paths:
  /requests/{reqID}:
    get:
      summary: Retrieve email request by reqID
      security:
        - bearerAuth: []
      parameters:
        - name: reqID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                type: object
                properties:
                  emails:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                        fulfilled:
                          type: string
                          nullable: true
                        lasterror:
                          type: string
                          nullable: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Unauthorized >:{"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "You've been banned from accessing service: {reason}"
        '503':
          description: Database is currently down, no request can be fulfilled at this time!
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Database is currently down, no request can be fulfilled at this time!"
  /requests/:
    post:
      summary: Submit an email request to queue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from:
                  type: string
                to:
                  type: array
                  items:
                    type: string
                replyto:
                  type: array
                  items:
                    type: string
                subject:
                  type: string
                text:
                  type: string
                html:
                  type: string
      responses:
        '200':
          description: Successful request
          content:
            application/json:
              schema:
                type: object
                properties:
                  reqID:
                    type: integer
                    default: 00000000-0000-0000-0000-000000000000
                  success:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Email successfully queued!"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Unauthorized >:{"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "You've been banned from accessing service: {reason}"
        '503':
          description: Database is currently down, no request can be fulfilled at this time!
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Database is currently down, no request can be fulfilled at this time!"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT